PHP Files
They may associate with some text files that you may need to create
manually (just a blank text file with the name specified in the php) if
they don't produce automatically. But other than they should work fine.
So I recommend just try them first and if they fail it means there is a
text file associated with it.