AlertZone
AlertZone now has kNN and Spatial query for 1 person. We can also detect
the people inside the zones. I don't know why it doesn't run after the
recent shutdown.
The only query that is left in AlertZone to be implemented is range
query to find the people within a range.
CrowdSourcing is in the beginning. Haven't pushed it yet.